feat(mcp): unify McpTarget on AgentProvider, add OpenCode/Pi/Hermes/Cursor install
- Add Cursor as a stub AgentProvider (no session parsing yet, MCP host only)
- Remove standalone McpTarget enum; reuse AgentProvider for all MCP targets
- Add install/uninstall/print_config for OpenCode (~/.config/opencode/opencode.json),
  Pi (~/.config/mcp/mcp.json), and Hermes (~/.hermes/config.yaml via serde_yaml)
- Fix Pi/OpenCode config path to use ~/.config (node_config_dir) instead of
  dirs::config_dir() which returns AppData\Roaming on Windows
- Include provider name in install/uninstall success messages
- Bump deps: ratatui 0.30, crossterm 0.29, dirs 6, getrandom 0.4, rusqlite 0.40,
  sha2 0.11, toml 1; adapt getrandom::fill, usize->i64 for rusqlite ToSql,
  ListState Copy for ratatui 0.30
- Add serde_yaml 0.9 dependency for Hermes config parsing

+++ crates/sivtr-core/src/history/query.rs
@@ -10,7 +10,7 @@ impl HistoryStore {
         )?;
 
         let entries = stmt
-            .query_map(rusqlite::params![limit], |row| {
+            .query_map(rusqlite::params![limit as i64], |row| {
                 Ok(HistoryEntry {
                     id: row.get(0)?,
                     content: row.get(1)?,
@@ -39,7 +39,7 @@ impl HistoryStore {
         )?;
 
         let entries = stmt
-            .query_map(rusqlite::params![fts_query, limit], |row| {
+            .query_map(rusqlite::params![fts_query, limit as i64], |row| {
                 Ok(HistoryEntry {
                     id: row.get(0)?,
                     content: row.get(1)?,
@@ -93,7 +93,7 @@ impl HistoryStore {
              WHERE id NOT IN (
                SELECT id FROM history ORDER BY id DESC LIMIT ?1
              )",
-            rusqlite::params![max_entries],
+            rusqlite::params![max_entries as i64],
         )?;
 
         Ok(())
